Make sure we use singlular in namespaces
Also did some cleanups.
