+++ src/Mailgun/Api/Suppression/Unsubscribe.php
@@ -7,13 +7,15 @@
  * of the MIT license. See the LICENSE file for details.
  */
 
-namespace Mailgun\Api;
+namespace Mailgun\Api\Supression;
 
+use Mailgun\Api\HttpApi;
+use Mailgun\Api\Pagination;
 use Mailgun\Assert;
-use Mailgun\Model\Suppressions\Unsubscribe\CreateResponse;
-use Mailgun\Model\Suppressions\Unsubscribe\DeleteResponse;
-use Mailgun\Model\Suppressions\Unsubscribe\IndexResponse;
-use Mailgun\Model\Suppressions\Unsubscribe\ShowResponse;
+use Mailgun\Model\Suppression\Unsubscribe\CreateResponse;
+use Mailgun\Model\Suppression\Unsubscribe\DeleteResponse;
+use Mailgun\Model\Suppression\Unsubscribe\IndexResponse;
+use Mailgun\Model\Suppression\Unsubscribe\ShowResponse;
 
 /**
  * @see https://documentation.mailgun.com/api-suppressions.html#unsubscribes
@@ -41,7 +43,7 @@ class Unsubscribe extends HttpApi
 
         $response = $this->httpGet(sprintf('/v3/%s/unsubscribes', $domain), $params);
 
-        return $this->safeDeserialize($response, IndexResponse::class);
+        return $this->safeHydrate($response, IndexResponse::class);
     }
 
     /**
@@ -57,7 +59,7 @@ class Unsubscribe extends HttpApi
 
         $response = $this->httpGet(sprintf('/v3/%s/unsubscribes/%s', $domain, $address));
 
-        return $this->safeDeserialize($response, ShowResponse::class);
+        return $this->safeHydrate($response, ShowResponse::class);
     }
 
     /**
@@ -76,7 +78,7 @@ class Unsubscribe extends HttpApi
 
         $response = $this->httpPost(sprintf('/v3/%s/unsubscribes', $domain), $params);
 
-        return $this->safeDeserialize($response, CreateResponse::class);
+        return $this->safeHydrate($response, CreateResponse::class);
     }
 
     /**
@@ -92,7 +94,7 @@ class Unsubscribe extends HttpApi
 
         $response = $this->httpDelete(sprintf('/v3/%s/unsubscribes/%s', $domain, $address));
 
-        return $this->safeDeserialize($response, DeleteResponse::class);
+        return $this->safeHydrate($response, DeleteResponse::class);
     }
 
     /**
@@ -106,6 +108,6 @@ class Unsubscribe extends HttpApi
 
         $response = $this->httpDelete(sprintf('/v3/%s/unsubscribes', $domain));
 
-        return $this->safeDeserialize($response, DeleteResponse::class);
+        return $this->safeHydrate($response, DeleteResponse::class);
     }
 }
