+++ src/Mailgun/Model/Suppression/Bounce/CreateResponse.php
@@ -7,13 +7,13 @@
  * of the MIT license. See the LICENSE file for details.
  */
 
-namespace Mailgun\Model\Suppressions\Complaint;
+namespace Mailgun\Model\Suppression\Bounce;
 
-use Mailgun\Model\Suppressions\BaseResponse;
+use Mailgun\Model\Suppression\BaseResponse;
 
 /**
  * @author Sean Johnson <sean@mailgun.com>
  */
-final class DeleteResponse extends BaseResponse
+final class CreateResponse extends BaseResponse
 {
 }
